requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-cloud-aiplatform
faiss-cpu
numpy
python-dotenv
pypdf
streamlit 
requests